requirements.txt
# EOC Database API Service - Python Dependencies

# FastAPI Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Oracle Database Driver (modern python-oracledb)
oracledb>=2.0.0

# Data validation and configuration
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Environment variables
python-dotenv>=1.0.0

# Logging
loguru>=0.7.2

# Retry logic and resilience
tenacity>=8.2.3

# CORS and security
python-multipart>=0.0.6

# HTTP client for health checks
httpx>=0.25.2

